claude_agent_course/module_05_mcp_protocol/projects/project_02_mcp_filesystem_agent/pyproject.toml
[project]
name = "mcp-filesystem-agent"
version = "0.1.0"
description = "Filesystem MCP client project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
